+++ FO-DICOM.Core/Network/DicomServerFactory.cs
@@ -20,13 +20,16 @@ namespace FellowOakDicom.Network
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="userState">Optional parameters</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <see cref="DicomServer{T}"/>, that starts listening for connections in the background.</returns>
         IDicomServer Create<T>(
             int port,
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
             ILogger logger = null,
-            object userState = null) where T : DicomService, IDicomServiceProvider;
+            object userState = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider;
         
         /// <summary>
         /// Creates a DICOM server object.
@@ -37,6 +40,8 @@ namespace FellowOakDicom.Network
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="userState">Optional parameters.</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <see cref="DicomServer{T}"/>, that starts listening for connections in the background.</returns>
         IDicomServer Create<T>(
             string ipAddress,
@@ -44,7 +49,8 @@ namespace FellowOakDicom.Network
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
             ILogger logger = null,
-            object userState = null) where T : DicomService, IDicomServiceProvider;
+            object userState = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider;
         
         /// <summary>
         /// Creates a DICOM server object.
@@ -53,10 +59,11 @@ namespace FellowOakDicom.Network
         /// <typeparam name="TServer">Concrete DICOM server type to be returned.</typeparam>
         /// <param name="ipAddress">IP address(es) to listen to. Value <code>null</code> applies default, IPv4Any.</param>
         /// <param name="port">Port to listen to.</param>
-        /// <param name="userState">Optional optional parameters.</param>
+        /// <param name="userState">Optional parameters</param>
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <typeparamref name="TServer"/>, that starts listening for connections in the background.</returns>
         IDicomServer Create<T, TServer>(
             string ipAddress,
@@ -64,7 +71,8 @@ namespace FellowOakDicom.Network
             object userState = null,
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
-            ILogger logger = null) where T : DicomService, IDicomServiceProvider where TServer : IDicomServer<T>;
+            ILogger logger = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider where TServer : IDicomServer<T>;
     }
 
     public static class DicomServerFactory
@@ -77,15 +85,18 @@ namespace FellowOakDicom.Network
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="userState">Optional parameters</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <see cref="DicomServer{T}"/>, that starts listening for connections in the background.</returns>
         public static IDicomServer Create<T>(
             int port,
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
             ILogger logger = null,
-            object userState = null) where T : DicomService, IDicomServiceProvider
+            object userState = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider
             => Setup.ServiceProvider
-            .GetRequiredService<IDicomServerFactory>().Create<T>(port, tlsAcceptor, fallbackEncoding, logger, userState);
+            .GetRequiredService<IDicomServerFactory>().Create<T>(port, tlsAcceptor, fallbackEncoding, logger, userState, configure);
 
         /// <summary>
         /// Creates a DICOM server object out of DI-container.
@@ -95,7 +106,9 @@ namespace FellowOakDicom.Network
         /// <param name="port">Port to listen to.</param>
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
+        /// <param name="userState">Optional parameters</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <see cref="DicomServer{T}"/>, that starts listening for connections in the background.</returns>
         public static IDicomServer Create<T>(
             string ipAddress,
@@ -103,9 +116,10 @@ namespace FellowOakDicom.Network
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
             ILogger logger = null,
-            object userState = null) where T : DicomService, IDicomServiceProvider
+            object userState = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider
             => Setup.ServiceProvider
-            .GetRequiredService<IDicomServerFactory>().Create<T>(ipAddress, port, tlsAcceptor, fallbackEncoding, logger, userState);
+            .GetRequiredService<IDicomServerFactory>().Create<T>(ipAddress, port, tlsAcceptor, fallbackEncoding, logger, userState, configure);
 
         /// <summary>
         /// Creates a DICOM server object out of DI-container.
@@ -118,6 +132,7 @@ namespace FellowOakDicom.Network
         /// <param name="tlsAcceptor">Handler to accept authenticated connections.</param>
         /// <param name="fallbackEncoding">Fallback encoding.</param>
         /// <param name="logger">Logger, if null default logger will be applied.</param>
+        /// <param name="configure">Configures the service options of the newly created DICOM server</param>
         /// <returns>An instance of <typeparamref name="TServer"/>, that starts listening for connections in the background.</returns>
         public static IDicomServer Create<T, TServer>(
             string ipAddress,
@@ -125,9 +140,10 @@ namespace FellowOakDicom.Network
             object userState = null,
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
-            ILogger logger = null) where T : DicomService, IDicomServiceProvider where TServer : IDicomServer<T>
+            ILogger logger = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider where TServer : IDicomServer<T>
             => Setup.ServiceProvider
-            .GetRequiredService<IDicomServerFactory>().Create<T, TServer>(ipAddress, port, userState, tlsAcceptor, fallbackEncoding, logger);
+            .GetRequiredService<IDicomServerFactory>().Create<T, TServer>(ipAddress, port, userState, tlsAcceptor, fallbackEncoding, logger, configure);
     }
 
     public class DefaultDicomServerFactory : IDicomServerFactory
@@ -135,15 +151,18 @@ namespace FellowOakDicom.Network
         private readonly IServiceScopeFactory  _serviceScopeFactory;
         private readonly IDicomServerRegistry _dicomServerRegistry;
         private readonly IOptions<DicomServiceOptions> _defaultServiceOptions;
+        private readonly IOptions<DicomServerOptions> _defaultServerOptions;
 
         public DefaultDicomServerFactory(
             IServiceScopeFactory  serviceScopeFactory,
             IDicomServerRegistry dicomServerRegistry,
-            IOptions<DicomServiceOptions> defaultServiceOptions)
+            IOptions<DicomServiceOptions> defaultServiceOptions, 
+            IOptions<DicomServerOptions> defaultServerOptions)
         {
             _serviceScopeFactory = serviceScopeFactory ?? throw new ArgumentNullException(nameof(serviceScopeFactory));
             _dicomServerRegistry = dicomServerRegistry ?? throw new ArgumentNullException(nameof(dicomServerRegistry));
             _defaultServiceOptions = defaultServiceOptions ?? throw new ArgumentNullException(nameof(defaultServiceOptions));
+            _defaultServerOptions = defaultServerOptions ?? throw new ArgumentNullException(nameof(defaultServerOptions));
         }
 
         public IDicomServer Create<T>(
@@ -151,12 +170,20 @@ namespace FellowOakDicom.Network
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null, 
             ILogger logger = null,
-            object userState = null)
+            object userState = null,
+            Action<DicomServerOptions> configure = null)
             where T : DicomService, IDicomServiceProvider 
-            => Create<T, DicomServer<T>>(NetworkManager.IPv4Any, port, userState, tlsAcceptor, fallbackEncoding, logger);
+            => Create<T, DicomServer<T>>(NetworkManager.IPv4Any, port, userState, tlsAcceptor, fallbackEncoding, logger, configure);
 
-        public IDicomServer Create<T>(string ipAddress, int port, ITlsAcceptor tlsAcceptor = null, Encoding fallbackEncoding = null, ILogger logger = null, object userState = null) where T : DicomService, IDicomServiceProvider
-            => Create<T, DicomServer<T>>(ipAddress, port, userState, tlsAcceptor, fallbackEncoding, logger);
+        public IDicomServer Create<T>(
+            string ipAddress,
+            int port, 
+            ITlsAcceptor tlsAcceptor = null,
+            Encoding fallbackEncoding = null,
+            ILogger logger = null, 
+            object userState = null,
+            Action<DicomServerOptions> configure = null) where T : DicomService, IDicomServiceProvider
+            => Create<T, DicomServer<T>>(ipAddress, port, userState, tlsAcceptor, fallbackEncoding, logger, configure);
 
         public virtual IDicomServer Create<TServiceProvider, TServer>(
             string ipAddress,
@@ -164,7 +191,8 @@ namespace FellowOakDicom.Network
             object userState = null,
             ITlsAcceptor tlsAcceptor = null,
             Encoding fallbackEncoding = null,
-            ILogger logger = null) where TServiceProvider : DicomService, IDicomServiceProvider where TServer : IDicomServer<TServiceProvider>
+            ILogger logger = null,
+            Action<DicomServerOptions> configure = null) where TServiceProvider : DicomService, IDicomServiceProvider where TServer : IDicomServer<TServiceProvider>
         {
             var dicomServerScope = _serviceScopeFactory.CreateScope();
             
@@ -183,11 +211,14 @@ namespace FellowOakDicom.Network
             server.ServiceScope = dicomServerScope;
 
             var serviceOptions = _defaultServiceOptions.Value.Clone();
+            var serverOptions = _defaultServerOptions.Value.Clone();
 
             // if not explicitly set, try to get a tls handler from DI container
             tlsAcceptor ??= dicomServerScope.ServiceProvider.GetService<ITlsAcceptor>();
-            
-            var runner = server.StartAsync(ipAddress, port, tlsAcceptor, fallbackEncoding, serviceOptions, userState);
+
+            configure?.Invoke(serverOptions);
+
+            var runner = server.StartAsync(ipAddress, port, tlsAcceptor, fallbackEncoding, serviceOptions, userState, serverOptions);
 
             if (server.Exception != null)
             {